[server] Refactor
